# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChefRole": {
        "enum": ["SOUS_CHEF", "HEAD_CHEF"],
        "type": "string",
        "x-shared-type": "ChefRoleValue"
      },
...
      "SystemRole": {
        "enum": ["ROOT_ADMIN"],
        "type": "string",
        "x-shared-type": "SystemRoleValue"
      },
      "User": {
        "additionalProperties": true,
        "properties": {
          "accountStatus": { "$ref": "#/components/schemas/UserAccountStatus" },
          "avatarUrl": { "nullable": true, "type": "string" },
          "chefRole": {
            "anyOf": [{ "$ref": "#/components/schemas/ChefRole" }],
            "nullable": true
          },
          "createdAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "name": { "type": "string" },
          "paymentRemindersEnabled": { "type": "boolean" },
          "phone": { "nullable": true, "type": "string" },
          "roles": { "items": { "type": "string" }, "type": "array" },
          "systemRole": {
            "anyOf": [{ "$ref": "#/components/schemas/SystemRole" }],
            "nullable": true
          },
          "username": { "type": "string" }
        },
        "required": [
          "id",
          "username",
          "name",
          "chefRole",
          "systemRole",
          "accountStatus",
          "roles"
        ],
        "type": "object"
      },
      "UserAccountStatus": {
        "enum": ["ACTIVE", "BLOCKED"],
        "type": "string",
        "x-shared-type": "UserAccountStatusValue"
      }
    },
    "securitySchemes": { "bearerAuth": { "scheme": "bearer", "type": "http" } }
  },
  "info": { "title": "FF RESTaurent API", "version": "1.1.0" },
  "openapi": "3.0.3",
  "paths": {
...
    "/admin/root-transfer": {
      "post": {
        "operationId": "postAdminRoot-transfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "confirmationUsername": {
                    "maxLength": 30,
                    "minLength": 3,
                    "type": "string"
                  },
                  "currentPassword": {
                    "maxLength": 128,
                    "minLength": 1,
                    "type": "string"
                  },
                  "targetUsername": {
                    "maxLength": 30,
                    "minLength": 3,
                    "type": "string"
                  }
                },
                "required": [
                  "currentPassword",
                  "targetUsername",
                  "confirmationUsername"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/members": {
      "get": {
        "operationId": "getMembers",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": { "maxLength": 100, "type": "string" }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": { "minLength": 1, "type": "string" }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "forward",
              "enum": ["forward", "backward"],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "name-asc",
              "enum": ["name-asc", "name-desc", "created-desc"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "items": {
                      "items": { "$ref": "#/components/schemas/User" },
                      "type": "array"
                    },
                    "pageInfo": {
                      "additionalProperties": true,
                      "properties": {
                        "endCursor": { "nullable": true, "type": "string" },
                        "hasNextPage": { "type": "boolean" },
                        "hasPreviousPage": { "type": "boolean" },
                        "startCursor": { "nullable": true, "type": "string" }
                      },
                      "required": ["endCursor", "hasNextPage"],
                      "type": "object"
                    }
                  },
                  "required": ["items", "pageInfo"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/users": {
      "get": {
        "operationId": "getUsers",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": { "maxLength": 100, "type": "string" }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": { "minLength": 1, "type": "string" }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "forward",
              "enum": ["forward", "backward"],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "name-asc",
              "enum": ["name-asc", "name-desc", "created-desc"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "items": {
                      "items": { "$ref": "#/components/schemas/User" },
                      "type": "array"
                    },
                    "pageInfo": {
                      "additionalProperties": true,
                      "properties": {
                        "endCursor": { "nullable": true, "type": "string" },
                        "hasNextPage": { "type": "boolean" },
                        "hasPreviousPage": { "type": "boolean" },
                        "startCursor": { "nullable": true, "type": "string" }
                      },
                      "required": ["endCursor", "hasNextPage"],
                      "type": "object"
                    }
                  },
                  "required": ["items", "pageInfo"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/users/{id}/account-status": {
      "patch": {
        "operationId": "patchUsersByIdAccount-status",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "minLength": 1, "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "accountStatus": {
                    "$ref": "#/components/schemas/UserAccountStatus"
                  }
                },
                "required": ["accountStatus"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            },
            "description": "Default Response"
          },
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/users/{id}/chef-role": {
      "patch": {
        "operationId": "patchUsersByIdChef-role",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "minLength": 1, "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "chefRole": {
                    "anyOf": [{ "$ref": "#/components/schemas/ChefRole" }],
                    "nullable": true
                  }
                },
                "required": ["chefRole"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            },
            "description": "Default Response"
          },
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    }
  }
}