# apps/api/src/routes/member-routes.ts
import type { FastifyInstance } from 'fastify';
import { ChefRole, SystemRole, UserAccountStatus } from '@prisma/client';
import {
  requireAuthenticatedUser,
  requireRootAdmin,
} from '../http/auth-guards.js';
import { prisma } from '../lib/prisma.js';
import { transferRootAdmin } from '../services/root-admin-service.js';
import { sanitizeUser } from '../lib/roles.js';
import { chefRoleSchema, rootAdminTransferSchema } from '../schemas/index.js';
import { memberQuerySchema } from '../schemas/index.js';
import { normalizeSearchQuery } from '../lib/search-normalization.js';
import { cursorPageResult } from '../lib/pagination.js';
import { updateUserAccountStatus } from '../services/user-account-service.js';
import { userAccountStatusSchema } from '../schemas/index.js';

const listMembers = async (
  queryValue: unknown,
  accountStatus?: UserAccountStatus,
) => {
  const query = memberQuerySchema.parse(queryValue);
  const orderBy =
    query.sort === 'name-desc'
      ? [{ name: 'desc' as const }, { id: 'desc' as const }]
      : query.sort === 'created-desc'
        ? [{ createdAt: 'desc' as const }, { id: 'desc' as const }]
        : [{ name: 'asc' as const }, { id: 'asc' as const }];
  const backward = query.direction === 'backward' && Boolean(query.cursor);
  const users = await prisma.user.findMany({
    where: {
      ...(accountStatus ? { accountStatus } : {}),
      ...(query.search
        ? { searchText: { contains: normalizeSearchQuery(query.search) } }
        : {}),
    },
    orderBy,
    ...(query.cursor ? { cursor: { id: query.cursor }, skip: 1 } : {}),
    take: backward ? -(query.limit + 1) : query.limit + 1,
  });
  return cursorPageResult(
    users.map(sanitizeUser),
    query.limit,
    backward,
    query.cursor,
  );
};

/**
 * Member routes keep regular member lookup separate from ROOT_ADMIN governance.
 */
export const registerMemberRoutes = (app: FastifyInstance) => {
  app.get(
    '/members',
    { preHandler: requireAuthenticatedUser },
    async (request) => listMembers(request.query, UserAccountStatus.ACTIVE),
  );

  app.get(
    '/users',
    { preHandler: [requireAuthenticatedUser, requireRootAdmin] },
    async (request) => listMembers(request.query),
  );

  app.patch(
    '/users/:id/account-status',
    { preHandler: [requireAuthenticatedUser, requireRootAdmin] },
    async (request) => {
      const { id } = request.params as { id: string };
      const body = userAccountStatusSchema.parse(request.body);
      const result = await updateUserAccountStatus({
        actorId: request.currentUser.id,
        targetId: id,
        accountStatus: body.accountStatus as UserAccountStatus,
      });
      if (result.changed) {
        request.log.warn({
          event: 'user_account_status_changed',
          actorId: request.currentUser.id,
          targetId: id,
          accountStatus: result.user.accountStatus,
        });
      }
      return sanitizeUser(result.user);
    },
  );

  app.patch(
    '/users/:id/chef-role',
    { preHandler: [requireAuthenticatedUser, requireRootAdmin] },
    async (request, reply) => {
      const { id } = request.params as { id: string };
      const body = chefRoleSchema.parse(request.body);
      const target = await prisma.user.findUnique({ where: { id } });
      if (!target) {
        return reply
          .code(404)
          .send({ code: 'NOT_FOUND', message: 'User not found' });
      }
      if (target.systemRole === SystemRole.ROOT_ADMIN) {
        return reply.code(403).send({
          code: 'ROOT_ADMIN_ROLE_CHANGE_FORBIDDEN',
          message: 'The ROOT_ADMIN cannot be changed through this endpoint',
        });
      }
      if (target.accountStatus === UserAccountStatus.BLOCKED) {
        return reply.code(400).send({
          code: 'BLOCKED_USER_ROLE_CHANGE_FORBIDDEN',
          message: 'Restore the account before changing its role',
        });
      }
      const updated = await prisma.$transaction(async (tx) => {
        const existing = await tx.user.findUniqueOrThrow({ where: { id } });
        const changedCount = await tx.user.updateMany({
          where: { id, systemRole: null },
          data: { chefRole: body.chefRole as ChefRole | null },
        });
        if (changedCount.count !== 1) return null;
        const changed = await tx.user.findUniqueOrThrow({ where: { id } });
        await tx.roleAuditLog.create({
          data: {
            userId: id,
            changedById: request.currentUser.id,
            fromRole: existing.chefRole,
            toRole: changed.chefRole,
          },
        });
        return changed;
      });
      if (!updated) {
        return reply.code(403).send({
          code: 'ROOT_ADMIN_ROLE_CHANGE_FORBIDDEN',
          message: 'The ROOT_ADMIN cannot be changed through this endpoint',
        });
      }
      return sanitizeUser(updated);
    },
  );

  app.post(
    '/admin/root-transfer',
    { preHandler: [requireAuthenticatedUser, requireRootAdmin] },
    async (request) => {
      const body = rootAdminTransferSchema.parse(request.body);
      const result = await transferRootAdmin({
        currentUserId: request.currentUser.id,
        ...body,
      });
      request.log.warn({
        event: 'root_admin_transferred',
        fromUserId: request.currentUser.id,
        toUserId: result.targetUserId,
        auditId: result.auditId,
      });
      return { ok: true };
    },
  );
};